+++ breakoutbox/Models/Domain/IGroepRepository.cs
@@ -2,7 +2,7 @@
 {
     public interface IGroepRepository
     {
-        Groep GetById(long ID);
+        Groep GetById(decimal ID);
         
     }
 }
